add a page for customizing the pipeline. rename the run_basic_client_ext.py to run_basic_client_ext.py
